backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6

# Two-Factor Authentication
pyotp==2.9.0
qrcode[pil]==7.4.2
pillow==10.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment and Configuration
python-dotenv==1.0.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Date and Time
python-dateutil==2.8.2

# Utilities
click==8.1.7
rich==13.7.0

# Development and Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Production dependencies
sentry-sdk[fastapi]==1.38.0
